+++ src/main/java/ru/schur/filmcatalogapp/service/UserService.java
@@ -3,6 +3,7 @@ package ru.schur.filmcatalogapp.service;
 import org.springframework.stereotype.Service;
 import ru.schur.filmcatalogapp.converter.UserConverter;
 import ru.schur.filmcatalogapp.dto.UserDTO;
+import ru.schur.filmcatalogapp.exception.ThereIsNoSuchUserException;
 import ru.schur.filmcatalogapp.model.User;
 import ru.schur.filmcatalogapp.repository.UserRepository;
 
@@ -28,7 +29,9 @@ public class UserService {
     }
 
     public User getUser(Long id) {
-        return userRepository.findById(id).orElseThrow(IllegalStateException::new);
+        return userRepository
+                .findById(id)
+                .orElseThrow(ThereIsNoSuchUserException::new);
     }
 
     public UserDTO saveUser(UserDTO userDTO) {
@@ -42,6 +45,8 @@ public class UserService {
     }
 
     public void deleteUserById(Long id) {
+        User user = getUser(id);
+        if(user == null) throw new ThereIsNoSuchUserException();
         userRepository.deleteById(id);
     }
 
@@ -51,6 +56,7 @@ public class UserService {
 
     public UserDTO editUser(Long id, UserDTO userDTO) {
         User user = getUser(id);
+        if(user == null) throw new ThereIsNoSuchUserException();
         user.setName(userDTO.getName());
         user.setAvatar(userDTO.getAvatar());
         user.setLogin(userDTO.getLogin());
@@ -60,11 +66,14 @@ public class UserService {
     }
 
     public List<UserDTO> findUserByName(String name) {
-        return userConverter.toUserDTOList(userRepository.findUserByName(name));
+        List<User> users = userRepository.findUserByName(name);
+        if(users.isEmpty()) throw new ThereIsNoSuchUserException();
+        return userConverter.toUserDTOList(users);
     }
 
-    public UserDTO addFilm(Long useId, Long filmId) {
-        User user = getUser(useId);
+    public UserDTO addFilm(Long userId, Long filmId) {
+        User user = getUser(userId);
+        if(user == null) throw new ThereIsNoSuchUserException();
         user.getFavouriteFilms().add(filmService.getFilm(filmId));
         return userConverter.toUserDTO(userRepository.save(user));
     }
